+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - `Polygon.FromAlignedBoundingBox2d(IEnumerable<Vector3>)`
 - `Random.NextColor()` and `Random.NextMaterial()`
 - `Validator.DisableValidationOnConstruction`
+- `Vector3.ComputeDefaultBasisVectors()`
 
 ### Changed
 - Make MeshConverter deserialization more flexible to accommodate a schema used in function `input_schema`.